Share make_position and make_tree test factories
